Add the connector and the manager classes
