+++ lib/event-bus/src/index.ts
@@ -1,3 +1,3 @@
 export * from './event-bus';
 export * from './mock-event-bus';
-export * from './rabbit-event-bus';
+export * from './rabbit-event-bus/amqp-manager';
