+++ lib/event-bus/src/rabbit-event-bus/types.ts
@@ -1,3 +1,4 @@
+import { EventIdentifier, Event } from '../event-bus';
 
 export interface StateChange<M extends object> {
   newState: 'CONNECTED' | 'NOT_CONNECTED' | 'NEW_MESSAGE';
@@ -11,4 +12,9 @@ export interface MessageWrapper<T> {
     retries: number;
     failures: number;
   };
+}
+
+export interface Subscription<P extends object> {
+  eventIdentifier: EventIdentifier;
+  handler: (ev: Event<P>) => Promise<boolean>;
 }
